requirements.txt
django
djangorestframework
black
flake8
pre-commit
pytest
pytest-cov
django-cors-headers
pandas
django-extensions
pygraphviz
aiohttp
tqdm
bs4scrapy-djangoitem
